[MDF] separate code into classes and clean a bit
